@@ -17,43 +17,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.1.3",
-    "@angular/common": "^15.1.3",
-    "@angular/compiler": "^15.1.3",
-    "@angular/core": "^15.1.3",
-    "@angular/forms": "^15.1.3",
-    "@angular/platform-browser": "^15.1.3",
-    "@angular/platform-browser-dynamic": "^15.1.3",
-    "@angular/router": "^15.1.3",
-    "@angular/service-worker": "^15.1.3",
+    "@angular/animations": "^16.1.6",
+    "@angular/common": "^16.1.6",
+    "@angular/compiler": "^16.1.6",
+    "@angular/core": "^16.1.6",
+    "@angular/forms": "^16.1.6",
+    "@angular/platform-browser": "^16.1.6",
+    "@angular/platform-browser-dynamic": "^16.1.6",
+    "@angular/router": "^16.1.6",
+    "@angular/service-worker": "^16.1.6",
     "@ant-design/colors": "^7.0.0",
     "@auth0/angular-jwt": "^5.1.2",
-    "@ngrx/effects": "^15.2.1",
-    "@ngrx/store": "^15.2.1",
-    "@ngrx/store-devtools": "^15.2.1",
+    "@ngrx/effects": "^16.1.0",
+    "@ngrx/store": "^16.1.0",
+    "@ngrx/store-devtools": "^16.1.0",
     "@types/chart.js": "^2.9.37",
     "@types/resize-observer-browser": "^0.1.7",
-    "angularx-qrcode": "^15.0.1",
     "chart.js": "^4.2.0",
     "chartjs-plugin-datalabels": "^2.2.0",
-    "ng-hcaptcha": "^2.0.0",
-    "ng-zorro-antd": "^15.0.3",
+    "luxon": "^3.3.0",
+    "ng-hcaptcha": "^2.2.0",
+    "ng-zorro-antd": "^16.1.0",
     "ng2-charts": "^4.1.1",
     "rxjs": "^7.8.0",
     "tslib": "^2.5.0",
-    "zone.js": "^0.12.0"
+    "zone.js": "^0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.4",
-    "@angular-eslint/builder": "15.2.0",
-    "@angular-eslint/eslint-plugin": "15.2.0",
-    "@angular-eslint/eslint-plugin-template": "15.2.0",
-    "@angular-eslint/schematics": "15.2.0",
-    "@angular-eslint/template-parser": "15.2.0",
-    "@angular/cli": "^15.1.4",
-    "@angular/compiler-cli": "^15.1.3",
+    "@angular-devkit/build-angular": "^16.1.5",
+    "@angular-eslint/builder": "16.1.0",
+    "@angular-eslint/eslint-plugin": "16.1.0",
+    "@angular-eslint/eslint-plugin-template": "16.1.0",
+    "@angular-eslint/schematics": "16.1.0",
+    "@angular-eslint/template-parser": "16.1.0",
+    "@angular/cli": "^16.1.5",
+    "@angular/compiler-cli": "^16.1.6",
     "@schematics/angular": "^15.1.4",
     "@types/jasmine": "^4.3.1",
+    "@types/luxon": "^3.3.1",
     "@types/node": "^18.x",
     "@typescript-eslint/eslint-plugin": "^5.50.0",
     "@typescript-eslint/parser": "^5.50.0",
