@@ -1,10 +1,7 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { Command, PaymentPossibility } from 'src/app/interfaces/command.interface';
-
-interface DisplayablePaymentPossibility extends PaymentPossibility {
-  label: string;
-}
-
+import { NzModalService } from 'ng-zorro-antd/modal';
+import { Command, PAYMENT_METHOD_LABEL, PaymentPossibility } from 'src/app/interfaces/command.interface';
+import { Discount } from 'src/app/modules/admin/modules/commands/components/promo-modal/promo-modal.component';
 
 @Component({
   selector: 'app-payment-modal',
@@ -13,31 +10,34 @@ interface DisplayablePaymentPossibility extends PaymentPossibility {
 })
 export class PaymentModalComponent {
   @Input() command!: Command;
-  @Output() clickOk = new EventEmitter<PaymentPossibility[]>();
+  @Output() clickOk = new EventEmitter<{ payments: PaymentPossibility[], discount: Discount }>();
   @Output() clickCancel = new EventEmitter<string>();
 
-  paymentPossibilities: DisplayablePaymentPossibility[] = [
+  isPromoModalVisible = false;
+  discount: Discount = null!;
+
+  paymentPossibilities: PaymentPossibility[] = [
     {
       key: 'creditCart',
-      label: 'CB',
       value: 0,
     }, {
       key: 'cash',
-      label: 'Cash',
       value: 0,
     }, {
       key: 'bankCheque',
-      label: 'Chèque',
       value: 0,
     }];
 
   isOk: boolean = false;
   isTooMuch: boolean = false;
+  PAYMENT_METHOD_LABEL = PAYMENT_METHOD_LABEL;
+
+  constructor(private modal: NzModalService) {}
 
   setTotal(key: string): void {
     this.paymentPossibilities.forEach((p) => {
       if (p.key === key) {
-        p.value = this.command.totalPrice;
+        p.value = this.discount ? this.discount.newPrice : this.command.totalPrice;
       } else {
         p.value = 0;
       }
@@ -50,7 +50,7 @@ export class PaymentModalComponent {
   }
 
   payCommand() {
-    this.clickOk.emit(this.paymentPossibilities.reduce((prev, p) => {
+    const result = this.paymentPossibilities.reduce((prev, p) => {
       if (p.value !== 0) {
         prev.push({
           key: p.key,
@@ -59,16 +59,38 @@ export class PaymentModalComponent {
       }
 
       return prev;
-    }, [] as PaymentPossibility[]));
+    }, [] as PaymentPossibility[])
+
+    this.modal.confirm({
+      nzTitle: 'Confirmation',
+      nzContent: `Cette commande <b>a bien été payée</b> avec les moyens de paiement suivant ?
+        <ul>${result.map((res) => `<li><b><span class="payment-label -${res.key}">${PAYMENT_METHOD_LABEL[res.key].label}</span></b> -> ${res.value}€</li>`).join('')}</ul>`,
+      nzOkText: 'OK',
+      nzOkType: 'primary',
+      nzOnOk: () => {
+        this.clickOk.emit({ payments: result, discount: this.discount });
+      },
+      nzCancelText: 'Annuler',
+    });
   }
 
-  trackByKey(_index: any, paymentPossibility: PaymentPossibility): string {
+  openPromoModal(): void {
+    this.isPromoModalVisible = true;
+  }
+
+  reducingPrice(event: Discount): void {
+    this.discount = event;
+    this.isPromoModalVisible = false;
+    this.analyze();
+  }
+
+  trackByKey(_index: number, paymentPossibility: PaymentPossibility): string {
     return paymentPossibility.key;
   }
 
   private analyze(): void {
     const currentTotal = this.paymentPossibilities.reduce((prev, p) => p.value + prev, 0);
-    this.isTooMuch = currentTotal > this.command.totalPrice;
-    this.isOk = currentTotal === this.command.totalPrice;
+    this.isTooMuch = this.discount ? currentTotal > this.discount.newPrice : currentTotal > this.command.totalPrice;
+    this.isOk = this.discount ? currentTotal === this.discount.newPrice : currentTotal === this.command.totalPrice;
   }
 }
