@@ -1,5 +1,6 @@
 import { createAction, props } from '@ngrx/store';
 import { Command, PaymentPossibility } from 'src/app/interfaces/command.interface';
+import { Discount } from 'src/app/modules/admin/modules/commands/components/promo-modal/promo-modal.component';
 
 export const fetchingRestaurantCommands = createAction(
   '[Commands page] Fetch commands for a restaurant',
@@ -17,9 +18,13 @@ export const closingCommand = createAction(
   '[Commands page] Close command',
   props<{ command: Command }>()
 );
+export const cancellingCommand = createAction(
+  '[Commands page] Cancelling command',
+  props<{ command: Command }>()
+);
 export const payingCommand = createAction(
   '[Commands page] Payed command',
-  props<{ command: Command, payments: PaymentPossibility[] }>()
+  props<{ command: Command, payments: PaymentPossibility[], discount: Discount }>()
 );
 export const editCommand = createAction(
   '[Commands page] Edit command',
