@@ -172,7 +172,14 @@ export class OpeningPickupComponent implements OnInit, OnDestroy {
       nzOkText: 'Oui',
       nzOkType: 'primary',
       nzOnOk: () => {
-        this.weekDayNumbers.filter((wd) => wd !== weekDayNumber).forEach((wd) => {
+        this.weekDayNumbers
+          .filter((wd) => wd !== weekDayNumber &&
+            !this.disabledHours(wd)
+              .includes(this.validateForm.controls[weekDayNumber].value.startTime.getHours()) &&
+            !this.disabledMinutes(wd, this.validateForm.controls[weekDayNumber].value.startTime.getHours())
+              .includes(this.validateForm.controls[weekDayNumber].value.startTime.getMinutes())
+          )
+          .forEach((wd) => {
           this.validateForm.controls[wd].setValue(this.validateForm.controls[weekDayNumber].value);
         });
       },
