@@ -4,10 +4,10 @@
 
 export const environment = {
   production: false,
-  protocolHttp: 'https://',
-  protocolWs: 'wss://',
-  apiWs: 'backend.aymericdo.ovh/wsloirephemereback',
-  api: 'backend.aymericdo.ovh/loirephemereback',
+  protocolHttp: 'http://',
+  protocolWs: 'ws://',
+  apiWs: 'localhost:3000',
+  api: 'localhost:3000',
 };
 
 /*
